+++ integration-tests/src/test_loop/tests/chunk_validator_kickout.rs
@@ -1,6 +1,6 @@
 use crate::test_loop::builder::TestLoopBuilder;
 use crate::test_loop::env::TestLoopEnv;
-use crate::test_loop::utils::client::get_epoch_all_validators;
+use crate::test_loop::utils::validators::get_epoch_all_validators;
 use crate::test_loop::utils::ONE_NEAR;
 use itertools::Itertools;
 use near_async::test_loop::data::TestLoopData;
