+++ integration-tests/src/test_loop/utils/mod.rs
@@ -1,7 +1,7 @@
-pub mod client;
 pub mod network;
 pub mod setups;
 pub mod transactions;
+pub mod validators;
 
 pub(crate) const ONE_NEAR: u128 = 1_000_000_000_000_000_000_000_000;
 pub(crate) const TGAS: u64 = 1_000_000_000_000;
